moved dtos to a shared project
