+++ DAPM/DAPM.Authenticator/DAPM.Authenticator.csproj
@@ -31,6 +31,11 @@
 
   <ItemGroup>
     <Folder Include="Migrations\" />
+    <Folder Include="Models\Dto\" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\UtilLibrary\UtilLibrary.csproj" />
   </ItemGroup>
 
 </Project>
