+++ DAPM/DAPM.sln
@@ -24,7 +24,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DAPM.Orchestrator", "DAPM.O
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DAPM.PeerApi", "DAPM.PeerApi\DAPM.PeerApi.csproj", "{33DD6ED0-1FA6-48BF-87CB-84D9BBE7AEAA}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DAPM.Authenticator", "DAPM.Authenticator\DAPM.Authenticator.csproj", "{1EC8D456-C8E0-44A4-B048-5AAD7A1F0240}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DAPM.Authenticator", "DAPM.Authenticator\DAPM.Authenticator.csproj", "{1EC8D456-C8E0-44A4-B048-5AAD7A1F0240}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UtilLibrary", "..\UtilLibrary\UtilLibrary.csproj", "{93BD44CA-B5BB-467E-9013-B0C70783CAD1}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -80,6 +82,10 @@ Global
 		{1EC8D456-C8E0-44A4-B048-5AAD7A1F0240}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1EC8D456-C8E0-44A4-B048-5AAD7A1F0240}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1EC8D456-C8E0-44A4-B048-5AAD7A1F0240}.Release|Any CPU.Build.0 = Release|Any CPU
+		{93BD44CA-B5BB-467E-9013-B0C70783CAD1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{93BD44CA-B5BB-467E-9013-B0C70783CAD1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{93BD44CA-B5BB-467E-9013-B0C70783CAD1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{93BD44CA-B5BB-467E-9013-B0C70783CAD1}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
