+++ UtilLibrary/OrgaizationDto.cs
@@ -1,6 +1,5 @@
-﻿using Microsoft.AspNetCore.Mvc;
-
-namespace DAPM.Authenticator.Models.Dto
+﻿
+namespace UtilLibrary
 {
     public class OrganisationsDto 
     {
